pyproject.toml
[tool.poetry]
name = "eva-mcp"
version = "1.0.0"
description = "Model Context Protocol (MCP) server for EVA Suite - exposes tools to AI agents via HTTP/SSE"
authors = ["EVA Suite Team <eva@example.com>"]
readme = "README.md"
packages = [{include = "eva_mcp", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
aiohttp = "^3.9.0"
azure-identity = "^1.15.0"
azure-keyvault-secrets = "^4.7.0"
azure-cosmos = "^4.5.0"
azure-mgmt-resource = "^23.0.0"
python-dotenv = "^1.0.0"
gitpython = "^3.1.40"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
httpx = "^0.25.0"
black = "^23.12.0"
ruff = "^0.1.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--verbose --strict-markers"

[tool.black]
line-length = 120
target-version = ['py311']

[tool.ruff]
line-length = 120
target-version = "py311"
